Winsockヘッダのushortとコンフリクトするのでポート番号の型をushortからsize_tに変更
Former-commit-id: d09882924cda0e77412e6ab05e784329398ee7db

+++ Nyx/Source/Network/WinsockTCPClientSocket.h
@@ -39,7 +39,7 @@ namespace Nyx {
 		* @param const char* アドレス
 		* @param const int ポート番号
 		*/
-		bool Connect(const char* address, ushort port);
+		bool Connect(const char* address, size_t port);
 		
 		
 		/**
@@ -47,7 +47,7 @@ namespace Nyx {
 		* @param const std::string& アドレス
 		* @param int ポート番号
 		*/
-		bool Connect(const std::string&  address, ushort port);
+		bool Connect(const std::string&  address, size_t port);
 		
 		
 		/**
